requirements.txt
prometheus-client==0.16.0
psutil==5.9.7
python-dotenv==1.0.0
